% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{ConvertPng}
\alias{ConvertPng}
\title{Convert Latex to PNG}
\usage{
ConvertPng(tex, dpi = 72)
}
\arguments{
\item{tex}{The name (path) of the TeX document.}

\item{dpi}{resolution (dots per inch) to render}
}
\description{
Convert standalone Latex to PNG.
}
\details{
Some TeX installations on Windows do not
have ‘texi2dvi.exe’. If ‘texify.exe’ is present
(e.g., part of MikTeX), then it can be used instead:
set \code{options(texi2dvi="texify.exe")}
or to the full path of the program. With TeX Live,
I use \code{options(texi2dvi="pdflatex.exe")}.
}

