% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikz.R
\name{TikzPng}
\alias{TikzPng}
\title{Tikz plot in PNG}
\usage{
TikzPng(fun.plot, name, width = 4, height = width/1.2, dpi = 72)
}
\arguments{
\item{fun.plot}{The function containing the code of the code you want.}

\item{name}{The base name of the PNG image to be created.}

\item{width}{The width of the output figure, in \bold{inches}.}

\item{height}{The height of the output figure, in \bold{inches}.}

\item{dpi}{resolution (dots per inch) to render}
}
\description{
Tikz plot in PNG.
}
\details{
Make sure you use escapes for special characters in Latex.
}

